requirements.txt
fastapi~=0.110.1
uvicorn[standard]~=0.29.0
python-dotenv~=1.0.1
loguru
sqlmodel~=0.0.16
pyjwt~=2.8.0
passlib~=1.7.4
python-multipart
alembic~=1.13.1
bcrypt
pydantic~=2.5.3
SQLAlchemy~=2.0.29
psycopg2-binary
